pyproject.toml
[project]
name = "ankihelper"
version = "0.0.0"
description = "A python package attempt to help you to create a Deck in term of Anki"
authors = [
  {name = "fugashy", email = "fugashy@icloud.com"}]
dependencies = [
  "accelerate",
  "click",
  "diffusers",
  "genanki",
  "googletrans==4.0.0-rc1",
  "google-cloud-translate==2.0.1",
  "gtts",
  "icecream",
  "matplotlib",
  "numpy",
  "pandas",
  "pydub",
  "scikit-learn",
  "scipy",
  "sentence-transformers",
  "torch",
  "torchaudio",
  "torchvision",
  "tqdm",
  "transformers",
  "whisper",
  "yt_dlp",
  "pdfplumber",
  ]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm]
python-version = 3.10


[project.scripts]
ankihelper = "ankihelper.cli:main"
